backend/requirements.txt
flask
flask-cors
torch
torchvision
pillow
pymongo
